+++ src/main/java/si/feri/um/repositories/RatingRepository.java
@@ -6,8 +6,8 @@ import si.feri.um.vao.Rating;
 import java.util.List;
 
 public interface RatingRepository extends CrudRepository<Rating, Integer> {
-    List<Rating> getRatingsByUserIdUser(Long idUser);
-    List<Rating> getRatingsByRestaurantIdRestaurant(Long idRestaurant);
+    List<Rating> getRatingsByUserIdUser(int idUser);
+    List<Rating> getRatingsByRestaurantIdRestaurant(int idRestaurant);
     List<Rating> getRatingsByUser_GoogleUserId(String googleUserId);
-    Rating getRatingByRestaurant_IdRestaurantAndUser_GoogleUserId(Long idRestaurant, String googleUserId);
+    Rating getRatingByRestaurant_IdRestaurantAndUser_GoogleUserId(int idRestaurant, String googleUserId);
 }
